requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
yfinance>=0.2
fredapi>=0.5
requests>=2.28
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
streamlit>=1.28
torch>=2.0
joblib>=1.3
scipy>=1.11
pyyaml>=6.0
tabulate
gdown>=5.0
ccxt>=4.0
